Add password recovery and verification views
Added new routes in AppShell.xaml for PasswordRecoveryPageView, VerifyCodePageView, NewPasswordPageView, HomePageView, and FoodRecipePageView. Included new XAML files in MAUIRecipeApp.csproj for PasswordVerifyCodePageView. Registered new views and view models in MauiProgram.cs for PasswordRecoveryPageView and PasswordRecoveryPageViewModel.

Enhanced LoginPageView.xaml with a TapGestureRecognizer for the "Sign Up" label to bind it to the SignUpCommand. Updated PasswordRecoveryPageView.xaml to set the DataType to PasswordRecoveryPageViewModel and bound the SendCodeCommand to the "Sign In" button. Modified PasswordRecoveryPageView.xaml.cs to accept a PasswordRecoveryPageViewModel in the constructor and set it as the BindingContext.

Added SignUpCommand in LoginPageViewModel.cs using the RelayCommand attribute to navigate to the "Sign Up" page. Created PasswordRecoveryPageViewModel.cs with a SendCodeCommand to navigate to the "Verify Code" page.

Added PasswordVerifyCodePageView.xaml with UI elements for entering and verifying a code, and PasswordVerifyCodePageView.xaml.cs with a default constructor.